initial beast Actor.type framework
