+++ modules/item/item.js
@@ -13,11 +13,10 @@ export class HMItem extends Item {
 
     prepareDerivedData() {
         super.prepareDerivedData();
-        const {data, type} = this.data;
-        const actorData = this.actor ? this.actor.data : null;
+        const {type} = this.data;
 
-        if (type === 'skill')  { this._prepSkillData(data, actorData);  } else
-        if (type === 'weapon') { this._prepWeaponData(data, actorData); }
+        if (type === 'skill')  { this._prepSkillData();  } else
+        if (type === 'weapon') { this._prepWeaponData(); }
     }
 
     _prepArmorData() {
@@ -65,7 +64,7 @@ export class HMItem extends Item {
                 hp += hpStack.pop();
             }
 
-            hpStack.push(parseInt(pTable[i].hp.value) || 0);
+            hpStack.push(parseInt(pTable[i].hp.value, 10) || 0);
             if (reroll) rerolled = true;
         }
         hp += Math.max(...hpStack);
@@ -73,8 +72,8 @@ export class HMItem extends Item {
 
         // grab the level data off the ptable
         const feature = data.features;
-        for (let i in feature) {
-            bonus[i] = feature[i] ? pTable[level][i].value || 0 : 0;
+        for (const j in feature) {
+            bonus[j] = feature[j] ? pTable[level][j].value || 0 : 0;
         }
 
         // Saves
@@ -94,20 +93,24 @@ export class HMItem extends Item {
         }
     }
 
-    _prepSkillData(data, actorData) {
-        if (!actorData) return;
+    _prepSkillData() {
+        if (!this.actor?.data) { return; }
 
-        const { bonus, relevant } = data;
-        if (data.universal && bonus.mastery.value === 0) {
-            const abilities = actorData.data.abilities.total;
-            const stack = [];
+        const actorData = this.actor.data;
+        const {bonus, relevant, universal} = this.data.data;
 
-            for (const key in relevant) {
-                if (relevant[key]) stack.push(abilities[key].value);
-            }
-            const value = Math.min(...stack);
-            bonus.stats = {value, 'literacy': value, 'verbal': value};
-        } else { delete bonus.stats; }
+        if (this.actor.type === 'character') {
+            if (universal && bonus.mastery.value === 0) {
+                const abilities = actorData.data.abilities.total;
+                const stack = [];
+
+                for (const key in relevant) {
+                    if (relevant[key]) { stack.push(abilities[key].value); }
+                }
+                const value = Math.min(...stack);
+                bonus.stats = {value, 'literacy': value, 'verbal': value};
+            } else { delete bonus.stats; }
+        }
 
         for (const key in bonus.total) {
             let sum = -bonus.total[key];
@@ -116,26 +119,29 @@ export class HMItem extends Item {
         }
     }
 
-    _prepWeaponData(itemData, actorData) {
-        if (!actorData) return;
+    _prepWeaponData() {
+        if (!this.actor?.data) { return; }
+        const actorData   = this.actor.data;
+        const itemData    = this.data.data;
 
-        const armors   = [];
-        const shields  = [];
-        const armor    = {};
-        const shield   = {};
-        const defItems = actorData.items.filter((a) => a.type === 'armor' &&
-                                                       a.data.data.state.equipped);
+        const isCharacter = actorData.type === 'character';
+        const armors      = [];
+        const shields     = [];
+        const armor       = {};
+        const shield      = {};
+        const defItems    = actorData.items.filter((a) => a.type === 'armor'
+                                                       && a.data.data.state.equipped);
 
         // Splitting armor and shields for now, so we can manage stances later.
         for (let i = 0; i < defItems.length; i++) {
             const defItem = defItems[i];
             const defData = defItem.data.data;
-            // Without having finer control over prepData order, we must foce a prep here.
+            // Without having finer control over prepData order, we must force a prep here.
             defItem.prepareData();
             defData.shield.checked ? shields.push(defItem) : armors.push(defItem);
         }
 
-        const bonus     = itemData.bonus;
+        const {bonus}   = itemData;
 
         const cclass    = {};
         const race      = {};
@@ -148,11 +154,11 @@ export class HMItem extends Item {
         const profTable = HMTABLES.weapons.noprof;
         const wSkill    = itemData.skill;
         const profItem  = actorData.items.find((a) => {
-            return a.type === "proficiency" && a.name === itemData.proficiency;
+            return a.type === 'proficiency' && a.name === itemData.proficiency;
         });
 
         let j = 0;
-        for (let key in bonus.total) {
+        for (const key in bonus.total) {
             const profBonus = profItem ? profItem.data.data.bonus?.[key] || 0
                                        : profTable.table[wSkill] * profTable.vector[j++];
             spec[key]   = profBonus || 0;
@@ -170,16 +176,21 @@ export class HMItem extends Item {
                 shield[key] = (shield[key] || 0) + (shieldData[key] || 0);
             }
         }
-        if (!Object.values(stats).every( a => a === 0)) { bonus.stats = stats; }
-        if (!Object.values(spec).every( a => a === 0)) { bonus.spec = spec; }
-        if (!Object.values(cclass).every( a => a === 0)) { bonus.class = cclass; }
-        if (!Object.values(race).every( a => a === 0)) { bonus.race = race; }
-        if (!Object.values(armor).every( a => a === 0)) { bonus.armor = armor; }
-        if (!Object.values(shield).every( a => a === 0)) { bonus.shield = shield; }
 
-        for (let key in bonus.total) {
+        // TODO: Build a new data.data.bonus rather than clean the old one.
+        Object.values(armor).every((a) => a === 0)  ? delete bonus.armor  : bonus.armor  = armor;
+        Object.values(shield).every((a) => a === 0) ? delete bonus.shield : bonus.shield = shield;
+
+        if (isCharacter) {
+            Object.values(stats).every((a) => a === 0)  ? delete bonus.stats : bonus.stats = stats;
+            Object.values(cclass).every((a) => a === 0) ? delete bonus.class : bonus.class = cclass;
+            Object.values(race).every((a) => a === 0)   ? delete bonus.race  : bonus.race  = race;
+            Object.values(spec).every((a) => a === 0)   ? delete bonus.spec  : bonus.spec  = spec;
+        }
+
+        for (const key in bonus.total) {
             let sum = -bonus.total[key];
-            for (let state in bonus) { sum += bonus[state][key]; }
+            for (const state in bonus) { sum += bonus[state][key]; }
             bonus.total[key] = sum;
         }
     }
