+++ templates/actor/beast-base.hbs
